:sparkles:: move websockets into seperate file
